requirements.txt
numpy
PyQt5==5.15.2
pyqtgraph
pyserial
seaborn==0.12.2